backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
pymongo==4.6.0
motor==3.3.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-generativeai==0.3.2
scholarly==1.7.11
SpeechRecognition==3.10.0
email-validator==2.1.0.post1
PyAudio==0.2.14
PyPDF2==3.0.1
pdfplumber==0.10.3
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
Pillow==10.1.0
matplotlib==3.8.2
networkx==3.2.1
plotly==5.17.0
kaleido==0.2.1
Werkzeug==2.3.7
pydub==0.25.1
pytesseract==0.3.10
WeasyPrint==61.2
cssselect2==0.7.0
tinycss2==1.2.1
cairosvg==2.7.1